Move ExtractorException to Runtime module
